+++ app/src/main/java/com/example/mycryptoapp/ui/fragments/cryptooverview/CryptoOverviewFragment.kt
@@ -1,4 +1,4 @@
-package com.example.mycryptoapp.ui.fragments.overview
+package com.example.mycryptoapp.ui.fragments.cryptooverview
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,15 +7,14 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.mycryptoapp.R
 
-
-class CryptoTransactionsFragment : Fragment() {
+class CryptoOverviewFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_crypto_transactions, container, false)
+        return inflater.inflate(R.layout.fragment_crypto_overview, container, false)
     }
 
 }
