+++ app/src/main/java/com/example/mycryptoapp/ui/fragments/transactions/CryptosTransactionsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.mycryptoapp.ui.fragments.overview
+package com.example.mycryptoapp.ui.fragments.transactions
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,14 +7,15 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.mycryptoapp.R
 
-class CryptoOverviewFragment : Fragment() {
+
+class CryptosTransactionsFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_crypto_overview, container, false)
+        return inflater.inflate(R.layout.fragment_crypto_transactions, container, false)
     }
 
 }
