+++ app/src/main/res/layout/fragment_crypto_overview.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragments.overview.CryptoOverviewFragment">
+    tools:context=".ui.fragments.cryptooverview.CryptoOverviewFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
